Stage 5: Added shadow projections

+++ src/COSIG_RayTracer/Objects/Triangle.cs
@@ -64,7 +64,7 @@ namespace COSIG_RayTracing_Parser__ConsoleApp_.Objects
                 vectors[0].Z - invertTransformedRay.Origin.Z, vectors[0].Z - vectors[2].Z, invertTransformedRay.Direction_Normalized.Z, 0,
                 0, 0, 0, 1);
             float B = matrix_AB.GetDeterminant() / A_Det;
-            if (B <= (float)-1.0E-6) return false;
+            if (B <= -1.0E-6f) return false;
 
             // Y
             Matrix4x4 matrix_AY = new Matrix4x4(
@@ -73,7 +73,7 @@ namespace COSIG_RayTracing_Parser__ConsoleApp_.Objects
                 vectors[0].Z - vectors[1].Z, vectors[0].Z - invertTransformedRay.Origin.Z, invertTransformedRay.Direction_Normalized.Z, 0,
                 0, 0, 0, 1);
             float Y = matrix_AY.GetDeterminant() / A_Det;
-            if (Y <= (float)-1.0E-6 || B + Y >= (1.0f + (float)1.0E-6)) return false;
+            if (Y <= -1.0E-6f || B + Y >= (1.0f + 1.0E-6f)) return false;
 
             Vector3 P = vectors[0] + B * (vectors[1] - vectors[0]) + Y * (vectors[2] - vectors[0]);
             Vector4 P4 = Transformation.TransformVector4(transformation.TransformationMatrix, new Vector4(P, 1.0f));
